requirements.txt
# Automatically generated by httpsn://github.com/damnever/pigar.

beautifulsoup4==4.9.3
cleantext==1.1.4
facenet-pytorch==2.5.3
gradio
ImageHash==4.3.1
imageio==2.31.1
matplotlib==3.7.2
networkx==2.8.8
nltk==3.5
numpy==1.23.1
opencv-python==4.8.0.74
google-cloud-vision
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
googletrans
pandas==2.0.3
Pillow==9.1.0
requests==2.31.0
scikit-learn==0.24.2
scipy==1.10.1
sentence-transformers==2.2.2
timm==0.6.11
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.35.2

